requirements.txt
python-dotenv
pandas
pydantic
google-generativeai
PyMuPDF
tqdm
